Rearrange code
