+++ src/Engine/Events/EventIR.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <unordered_map>
-#include <vector>
 #include <array>
 #include <string>
 
@@ -170,16 +168,3 @@ class EventIR {
     } data;
 };
 
-class EventMap {
- public:
-    void add(int id, EventIR ir);
-    void clear() { _eventsById.clear(); }
-
-    void dumpAll();
-    void dump(int id);
- private:
-    std::unordered_map<int, std::vector<EventIR>> _eventsById;
-};
-
-extern EventMap globalEventMap;
-extern EventMap localEventMap;
