--- src/Engine/Events/EventPrinter.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include "Engine/Events/EventParser.h"
